requirements.txt
fastapi
uvicorn
neo4j
openai
python-multipart
faster-whisper
sounddevice
scipy
requests
